Large refactor for logging: move into separate platform as a stand alone, add dependencies for some sample applications
