+++ examples/all-clusters-app/esp32/main/component.mk
@@ -44,6 +44,7 @@ COMPONENT_SRCDIRS :=                                                            
 
 COMPONENT_EXTRA_INCLUDES := $(PROJECT_PATH)/third_party/connectedhomeip/src/app/util                                   \
                             $(PROJECT_PATH)/third_party/connectedhomeip/src/app/server                                 \
+                            $(PROJECT_PATH)/third_party/connectedhomeip/src/platform/logging/include                   \
                             $(PROJECT_PATH)/third_party/connectedhomeip/examples/all-clusters-app/all-clusters-common  \
                             $(PROJECT_PATH)/third_party/connectedhomeip/third_party/nlio/repo/include                  \
                             $(PROJECT_PATH)/third_party/connectedhomeip/src
