+++ src/lib/core/CHIPConfig.h
@@ -72,21 +72,7 @@
 #include CHIP_PLATFORM_CONFIG_INCLUDE
 #endif
 
-// clang-format off
-
-
 // Profile-specific Configuration Headers
-/*
-#include "CHIPBDXConfig.h"
-
-#include "CHIPDMConfig.h"
-
-#include "CHIPTimeConfig.h"
-
-#include "CHIPEventLoggingConfig.h"
-
-#include "CHIPRMPConfig.h"
-*/
 /**
  *  @def CHIP_CONFIG_ERROR_TYPE
  *
@@ -97,7 +83,7 @@
 #ifndef CHIP_CONFIG_ERROR_TYPE
 #include <stdint.h>
 
-#define CHIP_CONFIG_ERROR_TYPE                             int32_t
+#define CHIP_CONFIG_ERROR_TYPE int32_t
 #endif // CHIP_CONFIG_ERROR_TYPE
 
 /**
@@ -108,7 +94,7 @@
  *
  */
 #ifndef CHIP_CONFIG_NO_ERROR
-#define CHIP_CONFIG_NO_ERROR                               0
+#define CHIP_CONFIG_NO_ERROR 0
 #endif // CHIP_CONFIG_NO_ERROR
 
 /**
@@ -119,7 +105,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ERROR_MIN
-#define CHIP_CONFIG_ERROR_MIN                              4000
+#define CHIP_CONFIG_ERROR_MIN 4000
 #endif // CHIP_CONFIG_ERROR_MIN
 
 /**
@@ -130,7 +116,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ERROR_MAX
-#define CHIP_CONFIG_ERROR_MAX                              4999
+#define CHIP_CONFIG_ERROR_MAX 4999
 #endif // CHIP_CONFIG_ERROR_MAX
 
 /**
@@ -142,7 +128,7 @@
  *
  */
 #ifndef _CHIP_CONFIG_ERROR
-#define _CHIP_CONFIG_ERROR(e)                              (CHIP_ERROR_MIN + (e))
+#define _CHIP_CONFIG_ERROR(e) (CHIP_ERROR_MIN + (e))
 #endif // _CHIP_CONFIG_ERROR
 
 /**
@@ -156,7 +142,7 @@
  *    #CHIP_CONFIG_USE_MICRO_ECC.
  */
 #ifndef CHIP_CONFIG_USE_OPENSSL_ECC
-#define CHIP_CONFIG_USE_OPENSSL_ECC                        1
+#define CHIP_CONFIG_USE_OPENSSL_ECC 1
 #endif // CHIP_CONFIG_USE_OPENSSL_ECC
 
 /**
@@ -171,7 +157,7 @@
  *
  */
 #ifndef CHIP_CONFIG_USE_MICRO_ECC
-#define CHIP_CONFIG_USE_MICRO_ECC                          0
+#define CHIP_CONFIG_USE_MICRO_ECC 0
 #endif // CHIP_CONFIG_USE_MICRO_ECC
 
 #if CHIP_CONFIG_USE_MICRO_ECC && CHIP_CONFIG_USE_OPENSSL_ECC
@@ -202,7 +188,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1       0
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1 0
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP160R1
 
 /**
@@ -214,7 +200,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1       1
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1 1
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1
 
 /**
@@ -227,7 +213,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1       1
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1 1
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1
 
 /**
@@ -241,7 +227,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
-#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1       1
+#define CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1 1
 #endif // CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
 
 /**
@@ -299,7 +285,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY         0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY
 
 /**
@@ -311,7 +297,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG1
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG1                   1
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG1 1
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG1
 
 /**
@@ -327,7 +313,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG2
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG2                   0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG2 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG2
 
 /**
@@ -343,7 +329,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG3
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG3                   0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG3 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG3
 
 /**
@@ -360,7 +346,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG4
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG4                   1
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG4 1
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG4
 
 /**
@@ -376,7 +362,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASE_CONFIG5
-#define CHIP_CONFIG_SUPPORT_PASE_CONFIG5                   0
+#define CHIP_CONFIG_SUPPORT_PASE_CONFIG5 0
 #endif // CHIP_CONFIG_SUPPORT_PASE_CONFIG5
 
 /**
@@ -412,9 +398,9 @@
  */
 #ifndef CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT
 #if CHIP_CONFIG_USE_MICRO_ECC
-#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT         1
+#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT 1
 #else
-#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT         0
+#define CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT 0
 #endif // CHIP_CONFIG_USE_MICRO_ECC
 #endif // CHIP_CONFIG_PASE_MESSAGE_PAYLOAD_ALIGNMENT
 
@@ -428,7 +414,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT
-#define CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT              15000
+#define CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT 15000
 #endif // CHIP_CONFIG_PASE_RATE_LIMITER_TIMEOUT
 
 /**
@@ -447,7 +433,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS
-#define CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS         3
+#define CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS 3
 #endif // CHIP_CONFIG_PASE_RATE_LIMITER_MAX_ATTEMPTS
 
 /**
@@ -481,7 +467,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_MGMT_PLATFORM
-#define CHIP_CONFIG_MEMORY_MGMT_PLATFORM      0
+#define CHIP_CONFIG_MEMORY_MGMT_PLATFORM 0
 #endif // CHIP_CONFIG_MEMORY_MGMT_PLATFORM
 
 /**
@@ -499,7 +485,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_MGMT_SIMPLE
-#define CHIP_CONFIG_MEMORY_MGMT_SIMPLE        0
+#define CHIP_CONFIG_MEMORY_MGMT_SIMPLE 0
 #endif // CHIP_CONFIG_MEMORY_MGMT_SIMPLE
 
 /**
@@ -517,7 +503,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MEMORY_MGMT_MALLOC
-#define CHIP_CONFIG_MEMORY_MGMT_MALLOC        1
+#define CHIP_CONFIG_MEMORY_MGMT_MALLOC 1
 #endif // CHIP_CONFIG_MEMORY_MGMT_MALLOC
 
 /**
@@ -525,7 +511,8 @@
  */
 
 #if ((CHIP_CONFIG_MEMORY_MGMT_PLATFORM + CHIP_CONFIG_MEMORY_MGMT_SIMPLE + CHIP_CONFIG_MEMORY_MGMT_MALLOC) != 1)
-#error "Please assert exactly one of CHIP_CONFIG_MEMORY_MGMT_PLATFORM, CHIP_CONFIG_MEMORY_MGMT_SIMPLE, or CHIP_CONFIG_MEMORY_MGMT_MALLOC."
+#error                                                                                                                             \
+    "Please assert exactly one of CHIP_CONFIG_MEMORY_MGMT_PLATFORM, CHIP_CONFIG_MEMORY_MGMT_SIMPLE, or CHIP_CONFIG_MEMORY_MGMT_MALLOC."
 #endif // ((CHIP_CONFIG_MEMORY_MGMT_PLATFORM + CHIP_CONFIG_MEMORY_MGMT_SIMPLE + CHIP_CONFIG_MEMORY_MGMT_MALLOC) != 1)
 
 #if !CHIP_CONFIG_MEMORY_MGMT_MALLOC && CHIP_SYSTEM_CONFIG_USE_BSD_IFADDRS
@@ -547,7 +534,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS
-#define CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS     0
+#define CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS 0
 #endif // CHIP_CONFIG_SIMPLE_ALLOCATOR_USE_SMALL_BUFFERS
 
 /**
@@ -579,7 +566,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY
-#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY         1
+#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY 1
 #endif // CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_DUMMY
 
 /**
@@ -595,7 +582,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM
-#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM      0
+#define CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM 0
 #endif // CHIP_CONFIG_SECURITY_MGR_TIME_ALERTS_PLATFORM
 
 /**
@@ -636,7 +623,7 @@
  *
  */
 #ifndef CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM
-#define CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM            0
+#define CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM 0
 #endif // CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM
 
 /**
@@ -654,7 +641,7 @@
  *
  */
 #ifndef CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG
-#define CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG            0
+#define CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG 0
 #endif // CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG
 
 /**
@@ -670,17 +657,19 @@
  *
  */
 #ifndef CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL
-#define CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL             1
+#define CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL 1
 #endif // CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL
 
 /**
  *  @}
  */
 
-#if ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG + CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
-#error "Please assert exactly one of CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM, CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG, or CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL."
-#endif // ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG + CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
-
+#if ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG +                                          \
+      CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
+#error                                                                                                                             \
+    "Please assert exactly one of CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM, CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG, or CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL."
+#endif // ((CHIP_CONFIG_RNG_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_RNG_IMPLEMENTATION_CHIPDRBG +
+       // CHIP_CONFIG_RNG_IMPLEMENTATION_OPENSSL) != 1)
 
 /**
  *  @def CHIP_CONFIG_DEV_RANDOM_DRBG_SEED
@@ -695,7 +684,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEV_RANDOM_DRBG_SEED
-#define CHIP_CONFIG_DEV_RANDOM_DRBG_SEED              0
+#define CHIP_CONFIG_DEV_RANDOM_DRBG_SEED 0
 #endif // CHIP_CONFIG_DEV_RANDOM_DRBG_SEED
 
 /**
@@ -708,10 +697,8 @@
  *
  */
 #ifndef CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME
-#define CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME                 "/dev/urandom"
+#define CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME "/dev/urandom"
 #endif // CHIP_CONFIG_DEV_RANDOM_DEVICE_NAME
-
-
 
 /**
  *  @name chip AES Block Cipher Algorithm Implementation Configuration.
@@ -742,7 +729,7 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM
-#define CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM            0
+#define CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM 0
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM
 
 /**
@@ -757,7 +744,7 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL
-#define CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL             1
+#define CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL 1
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL
 
 /**
@@ -772,7 +759,7 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_AESNI
-#define CHIP_CONFIG_AES_IMPLEMENTATION_AESNI              0
+#define CHIP_CONFIG_AES_IMPLEMENTATION_AESNI 0
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_AESNI
 
 /**
@@ -787,16 +774,14 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS
-#define CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS              0
+#define CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS 0
 #endif // CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS
 
 /**
  *  @}
  */
 
-#if ((CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM + \
-      CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL + \
-      CHIP_CONFIG_AES_IMPLEMENTATION_AESNI + \
+#if ((CHIP_CONFIG_AES_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_AES_IMPLEMENTATION_OPENSSL + CHIP_CONFIG_AES_IMPLEMENTATION_AESNI +    \
       CHIP_CONFIG_AES_IMPLEMENTATION_MBEDTLS) != 1)
 #error "Please assert exactly one CHIP_CONFIG_AES_IMPLEMENTATION_... option."
 #endif
@@ -813,9 +798,8 @@
  *
  */
 #ifndef CHIP_CONFIG_AES_USE_EXPANDED_KEY
-#define CHIP_CONFIG_AES_USE_EXPANDED_KEY                   0
+#define CHIP_CONFIG_AES_USE_EXPANDED_KEY 0
 #endif // CHIP_CONFIG_AES_USE_EXPANDED_KEY
-
 
 /**
  *  @name chip SHA1 and SHA256 Hash Algorithms Implementation Configuration.
@@ -847,7 +831,7 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM           0
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM 0
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM
 
 /**
@@ -864,7 +848,7 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT           0
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT 0
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT
 
 /**
@@ -879,7 +863,7 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL            1
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL 1
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL
 
 /**
@@ -894,20 +878,17 @@
  *
  */
 #ifndef CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS
-#define CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS            0
+#define CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS 0
 #endif // CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS
 
 /**
  *  @}
  */
 
-#if ((CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM + \
-      CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT + \
-      CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL  + \
-      CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS) != 1)
+#if ((CHIP_CONFIG_HASH_IMPLEMENTATION_PLATFORM + CHIP_CONFIG_HASH_IMPLEMENTATION_MINCRYPT +                                        \
+      CHIP_CONFIG_HASH_IMPLEMENTATION_OPENSSL + CHIP_CONFIG_HASH_IMPLEMENTATION_MBEDTLS) != 1)
 #error "Please assert exactly one CHIP_CONFIG_HASH_IMPLEMENTATION_... option."
 #endif
-
 
 /**
  *  @name chip key export protocol configuration.
@@ -942,7 +923,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1
-#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1            1
+#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1 1
 #endif // CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG1
 
 /**
@@ -954,13 +935,12 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2
-#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2            1
+#define CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2 1
 #endif // CHIP_CONFIG_SUPPORT_KEY_EXPORT_CONFIG2
 
 /**
  *  @}
  */
-
 
 /**
  *  @def CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES
@@ -971,7 +951,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES
-#define CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES     0
+#define CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES 0
 #endif // CHIP_CONFIG_ALLOW_NON_STANDARD_ELLIPTIC_CURVES
 
 /**
@@ -982,7 +962,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_EC_BITS
-#define CHIP_CONFIG_MAX_EC_BITS                            256
+#define CHIP_CONFIG_MAX_EC_BITS 256
 #endif // CHIP_CONFIG_MAX_EC_BITS
 
 /**
@@ -993,7 +973,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_RSA_BITS
-#define CHIP_CONFIG_MAX_RSA_BITS                           4096
+#define CHIP_CONFIG_MAX_RSA_BITS 4096
 #endif // CHIP_CONFIG_MAX_RSA_BITS
 
 /**
@@ -1005,7 +985,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_PEER_NODES
-#define CHIP_CONFIG_MAX_PEER_NODES                         128
+#define CHIP_CONFIG_MAX_PEER_NODES 128
 #endif // CHIP_CONFIG_MAX_PEER_NODES
 
 /**
@@ -1016,7 +996,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_CONNECTIONS
-#define CHIP_CONFIG_MAX_CONNECTIONS                        INET_CONFIG_NUM_TCP_ENDPOINTS
+#define CHIP_CONFIG_MAX_CONNECTIONS INET_CONFIG_NUM_TCP_ENDPOINTS
 #endif // CHIP_CONFIG_MAX_CONNECTIONS
 
 /**
@@ -1031,7 +1011,7 @@
  *    inbound and outbound connections.
  */
 #ifndef CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS
-#define CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS           (CHIP_CONFIG_MAX_CONNECTIONS * 4 / 5)
+#define CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS (CHIP_CONFIG_MAX_CONNECTIONS * 4 / 5)
 #endif // CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS
 
 /**
@@ -1047,7 +1027,7 @@
  *    #CHIP_CONFIG_MAX_INCOMING_TCP_CONNECTIONS.
  */
 #ifndef CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP
-#define CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP    2
+#define CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP 2
 #endif // CHIP_CONFIG_MAX_INCOMING_TCP_CON_FROM_SINGLE_IP
 
 /**
@@ -1058,7 +1038,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_SESSION_KEYS
-#define CHIP_CONFIG_MAX_SESSION_KEYS                       CHIP_CONFIG_MAX_CONNECTIONS
+#define CHIP_CONFIG_MAX_SESSION_KEYS CHIP_CONFIG_MAX_CONNECTIONS
 #endif // CHIP_CONFIG_MAX_SESSION_KEYS
 
 /**
@@ -1073,7 +1053,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS
-#define CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS             4
+#define CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS 4
 #endif // CHIP_CONFIG_MAX_APPLICATION_EPOCH_KEYS
 
 /**
@@ -1086,7 +1066,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_APPLICATION_GROUPS
-#define CHIP_CONFIG_MAX_APPLICATION_GROUPS                 8
+#define CHIP_CONFIG_MAX_APPLICATION_GROUPS 8
 #endif // CHIP_CONFIG_MAX_APPLICATION_GROUPS
 
 /**
@@ -1098,7 +1078,7 @@
  *
  */
 #ifndef CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC
-#define CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC         1
+#define CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC 1
 #endif // CHIP_CONFIG_USE_APP_GROUP_KEYS_FOR_MSG_ENC
 
 /**
@@ -1123,7 +1103,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS
-#define CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS            (CHIP_CONFIG_MAX_APPLICATION_GROUPS + 1)
+#define CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS (CHIP_CONFIG_MAX_APPLICATION_GROUPS + 1)
 #endif // CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS
 
 #if !(CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS > 0 && CHIP_CONFIG_MAX_CACHED_MSG_ENC_APP_KEYS < 256)
@@ -1170,7 +1150,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY
-#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY     0
+#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY 0
 #endif // CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY
 
 /**
@@ -1184,7 +1164,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2
-#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2               1
+#define CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2 1
 #endif // CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG2
 
 /**
@@ -1200,7 +1180,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT
-#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT  30000
+#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT 30000
 #endif // CHIP_CONFIG_DEFAULT_SECURITY_SESSION_ESTABLISHMENT_TIMEOUT
 
 /**
@@ -1213,7 +1193,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT
-#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT           15000
+#define CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT 15000
 #endif // CHIP_CONFIG_DEFAULT_SECURITY_SESSION_IDLE_TIMEOUT
 
 /**
@@ -1225,7 +1205,7 @@
  *
  */
 #ifndef CHIP_CONFIG_NUM_MESSAGE_BUFS
-#define CHIP_CONFIG_NUM_MESSAGE_BUFS                       16
+#define CHIP_CONFIG_NUM_MESSAGE_BUFS 16
 #endif // CHIP_CONFIG_NUM_MESSAGE_BUFS
 
 /**
@@ -1237,7 +1217,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS
-#define CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS       32
+#define CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS 32
 #endif // CHIP_CONFIG_MAX_UNSOLICITED_MESSAGE_HANDLERS
 
 /**
@@ -1248,9 +1228,8 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS
-#define CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS                  16
+#define CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS 16
 #endif // CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS
-
 
 /**
  *  @def CHIP_CONFIG_CONNECT_IP_ADDRS
@@ -1261,7 +1240,7 @@
  *
  */
 #ifndef CHIP_CONFIG_CONNECT_IP_ADDRS
-#define CHIP_CONFIG_CONNECT_IP_ADDRS                       4
+#define CHIP_CONFIG_CONNECT_IP_ADDRS 4
 #endif // CHIP_CONFIG_CONNECT_IP_ADDRS
 
 /**
@@ -1276,7 +1255,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE
-#define CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE                   1280
+#define CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE 1280
 #endif // CHIP_CONFIG_DEFAULT_UDP_MTU_SIZE
 
 /**
@@ -1310,7 +1289,7 @@
  *
  */
 #ifndef CHIP_HEADER_RESERVE_SIZE
-#define CHIP_HEADER_RESERVE_SIZE                           38
+#define CHIP_HEADER_RESERVE_SIZE 38
 #endif // CHIP_HEADER_RESERVE_SIZE
 
 /**
@@ -1321,7 +1300,7 @@
  *
  */
 #ifndef CHIP_TRAILER_RESERVE_SIZE
-#define CHIP_TRAILER_RESERVE_SIZE                          20
+#define CHIP_TRAILER_RESERVE_SIZE 20
 #endif // CHIP_TRAILER_RESERVE_SIZE
 
 /**
@@ -1332,7 +1311,7 @@
  *
  */
 #ifndef CHIP_PORT
-#define CHIP_PORT                                          11097
+#define CHIP_PORT 11097
 #endif // CHIP_PORT
 
 /**
@@ -1343,7 +1322,7 @@
  *
  */
 #ifndef CHIP_UNSECURED_PORT
-#define CHIP_UNSECURED_PORT                                11096
+#define CHIP_UNSECURED_PORT 11096
 #endif // CHIP_UNSECURED_PORT
 
 /**
@@ -1353,7 +1332,7 @@
  *    Enable use of an ephemeral UDP source port for locally initiated chip exchanges.
  */
 #ifndef CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT
-#define CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT              0
+#define CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT 0
 #endif // CHIP_CONFIG_ENABLE_EPHEMERAL_UDP_PORT
 
 /**
@@ -1367,7 +1346,7 @@
  *    including message encryption. Because of this it SHOULD NEVER BE ENABLED IN PRODUCTION BUILDS.
  */
 #ifndef CHIP_CONFIG_SECURITY_TEST_MODE
-#define CHIP_CONFIG_SECURITY_TEST_MODE                     0
+#define CHIP_CONFIG_SECURITY_TEST_MODE 0
 #endif // CHIP_CONFIG_SECURITY_TEST_MODE
 
 /**
@@ -1377,7 +1356,7 @@
  *    Enable support for resolving hostnames with a DNS resolver.
  */
 #ifndef CHIP_CONFIG_ENABLE_DNS_RESOLVER
-#define CHIP_CONFIG_ENABLE_DNS_RESOLVER                    (INET_CONFIG_ENABLE_DNS_RESOLVER)
+#define CHIP_CONFIG_ENABLE_DNS_RESOLVER (INET_CONFIG_ENABLE_DNS_RESOLVER)
 #endif // CHIP_CONFIG_ENABLE_DNS_RESOLVER
 
 /**
@@ -1392,7 +1371,7 @@
  *    change in the future to \c TRUE in all cases.
  */
 #ifndef CHIP_CONFIG_RESOLVE_IPADDR_LITERAL
-#define CHIP_CONFIG_RESOLVE_IPADDR_LITERAL                 (CHIP_SYSTEM_CONFIG_USE_SOCKETS)
+#define CHIP_CONFIG_RESOLVE_IPADDR_LITERAL (CHIP_SYSTEM_CONFIG_USE_SOCKETS)
 #endif // CHIP_CONFIG_RESOLVE_IPADDR_LITERAL
 
 /**
@@ -1406,7 +1385,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_TARGETED_LISTEN
-#define CHIP_CONFIG_ENABLE_TARGETED_LISTEN                 (!CHIP_SYSTEM_CONFIG_USE_LWIP)
+#define CHIP_CONFIG_ENABLE_TARGETED_LISTEN (!CHIP_SYSTEM_CONFIG_USE_LWIP)
 #endif // CHIP_CONFIG_ENABLE_TARGETED_LISTEN
 
 /**
@@ -1420,7 +1399,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN
-#define CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN            0
+#define CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN 0
 #endif // CHIP_CONFIG_ENABLE_UNSECURED_TCP_LISTEN
 
 /**
@@ -1431,7 +1410,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEBUG_CERT_VALIDATION
-#define CHIP_CONFIG_DEBUG_CERT_VALIDATION                  1
+#define CHIP_CONFIG_DEBUG_CERT_VALIDATION 1
 #endif // CHIP_CONFIG_DEBUG_CERT_VALIDATION
 
 /**
@@ -1442,7 +1421,7 @@
  *
  */
 #ifndef CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID
-#define CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID       (chip::Profiles::Security::kChipCurveId_prime256v1)
+#define CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID (chip::Profiles::Security::kChipCurveId_prime256v1)
 #endif // CHIP_CONFIG_OPERATIONAL_DEVICE_CERT_CURVE_ID
 
 /**
@@ -1462,7 +1441,7 @@
  *
  */
 #ifndef CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE
-#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE   0x1B9C
+#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE 0x1B9C
 #endif // CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_BEFORE
 
 /**
@@ -1482,7 +1461,7 @@
  *
  */
 #ifndef CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER
-#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER    0x6444
+#define CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER 0x6444
 #endif // CHIP_CONFIG_OP_DEVICE_CERT_VALID_DATE_NOT_AFTER
 
 /**
@@ -1493,7 +1472,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_PASE_INITIATOR
-#define CHIP_CONFIG_ENABLE_PASE_INITIATOR                  1
+#define CHIP_CONFIG_ENABLE_PASE_INITIATOR 1
 #endif // CHIP_CONFIG_ENABLE_PASE_INITIATOR
 
 /**
@@ -1505,7 +1484,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_PASE_RESPONDER
-#define CHIP_CONFIG_ENABLE_PASE_RESPONDER                  1
+#define CHIP_CONFIG_ENABLE_PASE_RESPONDER 1
 #endif // CHIP_CONFIG_ENABLE_PASE_RESPONDER
 
 /**
@@ -1516,7 +1495,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_CASE_INITIATOR
-#define CHIP_CONFIG_ENABLE_CASE_INITIATOR                  1
+#define CHIP_CONFIG_ENABLE_CASE_INITIATOR 1
 #endif // CHIP_CONFIG_ENABLE_CASE_INITIATOR
 
 /**
@@ -1527,7 +1506,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_CASE_RESPONDER
-#define CHIP_CONFIG_ENABLE_CASE_RESPONDER                  1
+#define CHIP_CONFIG_ENABLE_CASE_RESPONDER 1
 #endif // CHIP_CONFIG_ENABLE_CASE_RESPONDER
 
 /**
@@ -1540,7 +1519,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_CASE_CONFIG1
-#define CHIP_CONFIG_SUPPORT_CASE_CONFIG1                   1
+#define CHIP_CONFIG_SUPPORT_CASE_CONFIG1 1
 #endif // CHIP_CONFIG_SUPPORT_CASE_CONFIG1
 
 /**
@@ -1552,13 +1531,13 @@
  */
 #ifndef CHIP_CONFIG_DEFAULT_CASE_CURVE_ID
 #if CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_secp224r1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_secp224r1)
 #elif CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_prime256v1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_prime256v1)
 #elif CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP192R1
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_prime192v1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_prime192v1)
 #else
-#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID                  (chip::Profiles::Security::kChipCurveId_secp160r1)
+#define CHIP_CONFIG_DEFAULT_CASE_CURVE_ID (chip::Profiles::Security::kChipCurveId_secp160r1)
 #endif
 #endif // CHIP_CONFIG_DEFAULT_CASE_CURVE_ID
 
@@ -1566,14 +1545,16 @@
  *  @def CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES
  *
  *  @brief
- *    Default set of ECDH curves allowed to be used in a CASE session (initiating or responding), if not overridden by the application.
+ *    Default set of ECDH curves allowed to be used in a CASE session (initiating or responding), if not overridden by the
+ * application.
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES
 #if CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP224R1 || CHIP_CONFIG_SUPPORT_ELLIPTIC_CURVE_SECP256R1
-#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES            (chip::Profiles::Security::kChipCurveSet_secp224r1|chip::Profiles::Security::kChipCurveSet_prime256v1)
+#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES                                                                                    \
+    (chip::Profiles::Security::kChipCurveSet_secp224r1 | chip::Profiles::Security::kChipCurveSet_prime256v1)
 #else
-#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES            (chip::Profiles::Security::kChipCurveSet_All)
+#define CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES (chip::Profiles::Security::kChipCurveSet_All)
 #endif
 #endif // CHIP_CONFIG_DEFAULT_CASE_ALLOWED_CURVES
 
@@ -1596,7 +1577,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES
-#define CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES          10
+#define CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES 10
 #endif // CHIP_CONFIG_MAX_SHARED_SESSIONS_END_NODES
 
 /**
@@ -1608,7 +1589,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION
-#define CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION       10
+#define CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION 10
 #endif // CHIP_CONFIG_MAX_END_NODES_PER_SHARED_SESSION
 
 /**
@@ -1619,7 +1600,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_TAKE_INITIATOR
-#define CHIP_CONFIG_ENABLE_TAKE_INITIATOR                  0
+#define CHIP_CONFIG_ENABLE_TAKE_INITIATOR 0
 #endif // CHIP_CONFIG_ENABLE_TAKE_INITIATOR
 
 /**
@@ -1630,7 +1611,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_TAKE_RESPONDER
-#define CHIP_CONFIG_ENABLE_TAKE_RESPONDER                  0
+#define CHIP_CONFIG_ENABLE_TAKE_RESPONDER 0
 #endif // CHIP_CONFIG_ENABLE_TAKE_RESPONDER
 
 /**
@@ -1641,7 +1622,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR
-#define CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR            1
+#define CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR 1
 #endif // CHIP_CONFIG_ENABLE_KEY_EXPORT_INITIATOR
 
 /**
@@ -1652,7 +1633,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER
-#define CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER            1
+#define CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER 1
 #endif // CHIP_CONFIG_ENABLE_KEY_EXPORT_RESPONDER
 
 /**
@@ -1688,7 +1669,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH
-#define CHIP_CONFIG_REQUIRE_AUTH                           1
+#define CHIP_CONFIG_REQUIRE_AUTH 1
 #endif // CHIP_CONFIG_REQUIRE_AUTH
 
 /**
@@ -1705,7 +1686,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL
-#define CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL            CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_DEVICE_CONTROL
 
 /**
@@ -1722,7 +1703,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV
-#define CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV               CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_FABRIC_PROV
 
 /**
@@ -1739,7 +1720,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV
-#define CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV              CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_NETWORK_PROV
 
 /**
@@ -1756,7 +1737,7 @@
  *
  */
 #ifndef CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV
-#define CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV              CHIP_CONFIG_REQUIRE_AUTH
+#define CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV CHIP_CONFIG_REQUIRE_AUTH
 #endif // CHIP_CONFIG_REQUIRE_AUTH_SERVICE_PROV
 
 /**
@@ -1777,7 +1758,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT
-#define CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT     1
+#define CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT 1
 #endif // CHIP_CONFIG_ENABLE_PROVISIONING_BUNDLE_SUPPORT
 
 /**
@@ -1789,7 +1770,7 @@
  *
  */
 #ifndef CHIP_ERROR_LOGGING
-#define CHIP_ERROR_LOGGING                                 1
+#define CHIP_ERROR_LOGGING 1
 #endif // CHIP_ERROR_LOGGING
 
 /**
@@ -1801,7 +1782,7 @@
  *
  */
 #ifndef CHIP_PROGRESS_LOGGING
-#define CHIP_PROGRESS_LOGGING                              1
+#define CHIP_PROGRESS_LOGGING 1
 #endif // CHIP_PROGRESS_LOGGING
 
 /**
@@ -1813,7 +1794,7 @@
  *
  */
 #ifndef CHIP_DETAIL_LOGGING
-#define CHIP_DETAIL_LOGGING                                1
+#define CHIP_DETAIL_LOGGING 1
 #endif // CHIP_DETAIL_LOGGING
 
 /**
@@ -1827,6 +1808,14 @@
  *
  */
 
+/**                                                                                                         │    │
+ * CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE
+ *
+ * The maximum size (in bytes) of a log message
+ */
+#ifndef CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE
+#define CHIP_CONFIG_LOG_MESSAGE_MAX_SIZE 256
+#endif
 
 /**
  *  @def CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING
@@ -1838,7 +1827,6 @@
 #ifndef CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING
 #define CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING 0
 #endif // CHIP_CONFIG_ENABLE_FUNCT_ERROR_LOGGING
-
 
 /**
  *  @def CHIP_CONFIG_ENABLE_CONDITION_LOGGING
@@ -1861,7 +1849,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS
-#define CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS      (10000)
+#define CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS (10000)
 #endif // CHIP_CONFIG_SERVICE_DIR_CONNECT_TIMEOUT_MSECS
 
 /**
@@ -1878,7 +1866,7 @@
  *
  */
 #ifndef CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT
-#define CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT        15000
+#define CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT 15000
 #endif // CHIP_CONFIG_DEFAULT_INCOMING_CONNECTION_IDLE_TIMEOUT
 
 /**
@@ -1892,7 +1880,7 @@
  *
  */
 #ifndef CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT
-#define CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT          2000
+#define CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT 2000
 #endif // CHIP_CONFIG_MSG_COUNTER_SYNC_RESP_TIMEOUT
 
 /**
@@ -1904,7 +1892,7 @@
  *
  */
 #ifndef CHIP_CONFIG_TEST
-#define CHIP_CONFIG_TEST                                   0
+#define CHIP_CONFIG_TEST 0
 #endif // CHIP_CONFIG_TEST
 
 /**
@@ -1916,7 +1904,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SHORT_ERROR_STR
-#define CHIP_CONFIG_SHORT_ERROR_STR                        0
+#define CHIP_CONFIG_SHORT_ERROR_STR 0
 #endif // CHIP_CONFIG_SHORT_ERROR_STR
 
 /**
@@ -1942,9 +1930,9 @@
  */
 #ifndef CHIP_CONFIG_ERROR_STR_SIZE
 #if CHIP_CONFIG_SHORT_ERROR_STR
-#define CHIP_CONFIG_ERROR_STR_SIZE                         (5 + 1 + 10 + 3 + 8 + 1)
+#define CHIP_CONFIG_ERROR_STR_SIZE (5 + 1 + 10 + 3 + 8 + 1)
 #else // CHIP_CONFIG_SHORT_ERROR_STR
-#define CHIP_CONFIG_ERROR_STR_SIZE                         256
+#define CHIP_CONFIG_ERROR_STR_SIZE 256
 #endif // CHIP_CONFIG_SHORT_ERROR_STR
 #endif // CHIP_CONFIG_ERROR_STR_SIZE
 
@@ -1957,7 +1945,7 @@
  *
  */
 #ifndef CHIP_CONFIG_CUSTOM_ERROR_FORMATTER
-#define CHIP_CONFIG_CUSTOM_ERROR_FORMATTER                 0
+#define CHIP_CONFIG_CUSTOM_ERROR_FORMATTER 0
 #endif // CHIP_CONFIG_CUSTOM_ERROR_FORMATTER
 
 /**
@@ -1975,7 +1963,7 @@
  *  and decimal values
  */
 #ifndef CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT
-#define CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT          "0x%08" PRIX32
+#define CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT "0x%08" PRIX32
 #endif // CHIP_CONFIG_SHORT_FORM_ERROR_VALUE_FORMAT
 
 /**
@@ -1988,7 +1976,7 @@
  *
  */
 #ifndef CHIP_CONFIG_BLE_PKT_RESERVED_SIZE
-#define CHIP_CONFIG_BLE_PKT_RESERVED_SIZE                  0
+#define CHIP_CONFIG_BLE_PKT_RESERVED_SIZE 0
 #endif // CHIP_CONFIG_BLE_PKT_RESERVED_SIZE
 
 /**
@@ -2000,7 +1988,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS
-#define CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS            1
+#define CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS 1
 #endif // CHIP_CONFIG_ENABLE_SECURITY_DEBUG_FUNCS
 
 /**
@@ -2027,26 +2015,22 @@
  */
 #ifndef CHIP_CONFIG_IsPlatformErrorNonCritical
 #if CHIP_SYSTEM_CONFIG_USE_LWIP
-#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE)         \
-    ((CODE) == chip::System::MapErrorLwIP(ERR_RTE)         || \
-     (CODE) == chip::System::MapErrorLwIP(ERR_MEM))
+#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE)                                                                          \
+    ((CODE) == chip::System::MapErrorLwIP(ERR_RTE) || (CODE) == chip::System::MapErrorLwIP(ERR_MEM))
 #else // !CHIP_SYSTEM_CONFIG_USE_LWIP
-#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE)  0
+#define _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE) 0
 #endif // !CHIP_SYSTEM_CONFIG_USE_LWIP
 
 #if CHIP_SYSTEM_CONFIG_USE_SOCKETS
-#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)        \
-    ((CODE) == chip::System::MapErrorPOSIX(EHOSTUNREACH)   || \
-     (CODE) == chip::System::MapErrorPOSIX(ENETUNREACH)    || \
-     (CODE) == chip::System::MapErrorPOSIX(EADDRNOTAVAIL)  || \
-     (CODE) == chip::System::MapErrorPOSIX(EPIPE))
+#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)                                                                         \
+    ((CODE) == chip::System::MapErrorPOSIX(EHOSTUNREACH) || (CODE) == chip::System::MapErrorPOSIX(ENETUNREACH) ||                  \
+     (CODE) == chip::System::MapErrorPOSIX(EADDRNOTAVAIL) || (CODE) == chip::System::MapErrorPOSIX(EPIPE))
 #else // !CHIP_SYSTEM_CONFIG_USE_SOCKETS
-#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)  0
+#define _CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE) 0
 #endif // !CHIP_SYSTEM_CONFIG_USE_SOCKETS
 
-#define CHIP_CONFIG_IsPlatformErrorNonCritical(CODE)              \
-    (_CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE)        || \
-     _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE))
+#define CHIP_CONFIG_IsPlatformErrorNonCritical(CODE)                                                                               \
+    (_CHIP_CONFIG_IsPlatformPOSIXErrorNonCritical(CODE) || _CHIP_CONFIG_IsPlatformLwIPErrorNonCritical(CODE))
 #endif // CHIP_CONFIG_IsPlatformErrorNonCritical
 
 /**
@@ -2059,7 +2043,7 @@
  *
  */
 #ifndef CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS
-#define CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS   0
+#define CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS 0
 #endif // CHIP_CONFIG_WILL_OVERRIDE_PLATFORM_MATH_FUNCS
 
 /**
@@ -2116,7 +2100,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID
-#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID      "EncMsgCntr"
+#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID "EncMsgCntr"
 #endif // CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_ID
 
 /**
@@ -2127,7 +2111,7 @@
  *
  */
 #ifndef CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH
-#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH   0x1000
+#define CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH 0x1000
 #endif // CHIP_CONFIG_PERSISTED_STORAGE_ENC_MSG_CNTR_EPOCH
 
 /**
@@ -2211,7 +2195,7 @@
  *
  */
 #ifndef CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE
-#define CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE     1
+#define CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE 1
 #endif // CHIP_CONFIG_SUPPORT_LEGACY_ADD_NETWORK_MESSAGE
 
 /**
@@ -2228,7 +2212,7 @@
  *
  */
 #ifndef CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE
-#define CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE     0
+#define CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE 0
 #endif // CHIP_CONFIG_ALWAYS_USE_LEGACY_ADD_NETWORK_MESSAGE
 
 /**
@@ -2239,7 +2223,7 @@
  * chip fabric.
  */
 #ifndef CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN
-#define CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN         0
+#define CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN 0
 #endif // CHIP_CONFIG_ENABLE_IFJ_SERVICE_FABRIC_JOIN
 
 /**
@@ -2250,7 +2234,7 @@
  * device connections across all supported transports.
  */
 #ifndef CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE
-#define CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE                   16
+#define CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE 16
 #endif // CHIP_CONFIG_PEER_CONNECTION_POOL_SIZE
 
 /**
@@ -2260,7 +2244,7 @@
  * expired.
  */
 #ifndef CHIP_PEER_CONNECTION_TIMEOUT_MS
-#define CHIP_PEER_CONNECTION_TIMEOUT_MS      120000
+#define CHIP_PEER_CONNECTION_TIMEOUT_MS 120000
 #endif // CHIP_PEER_CONNECTION_TIMEOUT_MS
 
 /**
@@ -2269,48 +2253,45 @@
  * @brief How frequent are peer connections checked for timeouts.
  */
 #ifndef CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS
-#define CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS      5000
+#define CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS 5000
 #endif // CHIP_PEER_CONNECTION_TIMEOUT_CHECK_FREQUENCY_MS
 
 /**
-   *  @def CHIP_CONFIG_MAX_BINDINGS
-   *
-   *  @brief
-   *    Maximum number of simultaneously active bindings per WeaveExchangeManager
-   *    The new single source TimeSync client takes one binding.
-   *    Every WDM one-way subscription takes one binding. Mutual subscription counts as two one-way subscriptions.
-   *    A reserved slot is needed to take an incoming subscription request.
-   *    For a device with 2 mutual subscriptions, and one single source time sync client, it needs 2 x 2 + 1 = 5 bindings at least.
-   *    At least six is needed if it still wants to take new WDM subscriptions under this load.
-   */
-  #ifndef CHIP_CONFIG_MAX_BINDINGS
-  #define CHIP_CONFIG_MAX_BINDINGS                           6
-  #endif // CHIP_CONFIG_MAX_BINDINGS
+ *  @def CHIP_CONFIG_MAX_BINDINGS
+ *
+ *  @brief
+ *    Maximum number of simultaneously active bindings per WeaveExchangeManager
+ *    The new single source TimeSync client takes one binding.
+ *    Every WDM one-way subscription takes one binding. Mutual subscription counts as two one-way subscriptions.
+ *    A reserved slot is needed to take an incoming subscription request.
+ *    For a device with 2 mutual subscriptions, and one single source time sync client, it needs 2 x 2 + 1 = 5 bindings at least.
+ *    At least six is needed if it still wants to take new WDM subscriptions under this load.
+ */
+#ifndef CHIP_CONFIG_MAX_BINDINGS
+#define CHIP_CONFIG_MAX_BINDINGS 6
+#endif // CHIP_CONFIG_MAX_BINDINGS
 
 /**
-   *  @def CHIP_CONFIG_MAX_DEVICE_ADMINS
-   *
-   *  @brief
-   *    Maximum number of administrators that can provision the device. Each admin
-   *    can provision the device with their unique operational credentials and manage
-   *    their access control lists.
-   */
-  #ifndef CHIP_CONFIG_MAX_DEVICE_ADMINS
-  #define CHIP_CONFIG_MAX_DEVICE_ADMINS                      16
-  #endif // CHIP_CONFIG_MAX_DEVICE_ADMINS
+ *  @def CHIP_CONFIG_MAX_DEVICE_ADMINS
+ *
+ *  @brief
+ *    Maximum number of administrators that can provision the device. Each admin
+ *    can provision the device with their unique operational credentials and manage
+ *    their access control lists.
+ */
+#ifndef CHIP_CONFIG_MAX_DEVICE_ADMINS
+#define CHIP_CONFIG_MAX_DEVICE_ADMINS 16
+#endif // CHIP_CONFIG_MAX_DEVICE_ADMINS
 
-  /**
+/**
  * @def CHIP_NON_PRODUCTION_MARKER
  *
  * @brief Defines the name of a mark symbol whose presence signals that the chip code
  * includes development/testing features that should never be used in production contexts.
  */
 #ifndef CHIP_NON_PRODUCTION_MARKER
-#if (CHIP_CONFIG_SECURITY_TEST_MODE || \
-     CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY || \
-     CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY || \
-     (!CHIP_CONFIG_REQUIRE_AUTH) || \
-     CHIP_FUZZING_ENABLED)
+#if (CHIP_CONFIG_SECURITY_TEST_MODE || CHIP_CONFIG_SUPPORT_PASE_CONFIG0_TEST_ONLY ||                                               \
+     CHIP_CONFIG_SUPPORT_PASSCODE_CONFIG1_TEST_ONLY || (!CHIP_CONFIG_REQUIRE_AUTH) || CHIP_FUZZING_ENABLED)
 #define CHIP_NON_PRODUCTION_MARKER WARNING__DO_NOT_SHIP__CONTAINS_NON_PRODUCTION_CHIP_CODE
 #endif
 #endif
@@ -2318,5 +2299,3 @@
 #ifdef CHIP_NON_PRODUCTION_MARKER
 extern const char CHIP_NON_PRODUCTION_MARKER[];
 #endif
-
-// clang-format on
