+++ src/lib/support/logging/CHIPLogging.h
@@ -36,6 +36,10 @@
 
 #include <core/CHIPConfig.h>
 
+#include <logging/LogV.h>
+
+#include "Constants.h"
+
 #include <stdarg.h>
 #include <stdint.h>
 
@@ -63,157 +67,9 @@
 namespace chip {
 namespace Logging {
 
-/**
- *  @enum LogModule
- *
- *  @brief
- *    Identifies a logical section of code that is a source of log
- *    messages.
- *
- *  @note If you add modules or rearrange this list you must update the
- *        ModuleNames tables in ChipLogging.cpp.
- *
- */
-enum LogModule
-{
-    kLogModule_NotSpecified = 0,
-
-    kLogModule_Inet,
-    kLogModule_Ble,
-    kLogModule_MessageLayer,
-    kLogModule_SecurityManager,
-    kLogModule_ExchangeManager,
-    kLogModule_TLV,
-    kLogModule_ASN1,
-    kLogModule_Crypto,
-    kLogModule_Controller,
-    kLogModule_Alarm,
-    kLogModule_BDX,
-    kLogModule_DataManagement,
-    kLogModule_DeviceControl,
-    kLogModule_DeviceDescription,
-    kLogModule_Echo,
-    kLogModule_FabricProvisioning,
-    kLogModule_NetworkProvisioning,
-    kLogModule_ServiceDirectory,
-    kLogModule_ServiceProvisioning,
-    kLogModule_SoftwareUpdate,
-    kLogModule_TokenPairing,
-    kLogModule_TimeService,
-    kLogModule_Heartbeat,
-    kLogModule_chipSystemLayer,
-    kLogModule_EventLogging,
-    kLogModule_Support,
-    kLogModule_chipTool,
-    kLogModule_Zcl,
-    kLogModule_Shell,
-    kLogModule_DeviceLayer,
-    kLogModule_SetupPayload,
-    kLogModule_AppServer,
-    kLogModule_Discovery,
-
-    kLogModule_Max
-};
-
-/**
- *  @enum LogCategory
- *
- *  @brief
- *    Identifies a category to which an particular error message
- *    belongs.
- *
- */
-enum LogCategory
-{
-    /*!<
-     *   This log category indicates, when passed to SetLogFilter(),
-     *   that no messages should be logged.
-     *
-     */
-    kLogCategory_None = 0,
-
-    /*!<
-     *   Indicates a category of log message that describes an unexpected
-     *   or severe failure.
-     *
-     *   This log category indicates that a logged message describes
-     *   an unexpected or severe failure in the code.
-     *
-     *   It should be used for things such as out-of-resource errors,
-     *   internal inconsistencies, API misuse, etc. In general, errors
-     *   that are expected to occur as part of normal operation, or
-     *   that are largely determined by external factors (e.g. network
-     *   errors, user/operator induced errors, etc.) should be logged
-     *   as kLogCategory_Progress messages, not as kLogCategory_Error
-     *   messages.
-     *
-     */
-    kLogCategory_Error = 1,
-
-    /*!<
-     *   Indicates a category of log message that describes an event
-     *   that marks the start or end of a major activity, or a major
-     *   change in the state of the overall system.
-     *
-     *   It should be reserved for high-level events. Such messages
-     *   should provide the log reader with a good sense of the
-     *   overall activity of the system at any point in time, while
-     *   being minimally verbose. Where necessary such messages should
-     *   include identifiers or other values that can be used to
-     *   correlate messages involving a common actor or subject
-     *   (e.g. connection ids, request ids, etc.) and/or to identify
-     *   types of actions being taken or handled (e.g.  message types,
-     *   requested resource types, error numbers, etc.).
-     *
-     */
-    kLogCategory_Progress = 2,
-
-    /*!<
-     *   Indicates a category of log message that describes detailed
-     *   information about an event or the state of the system.
-     *
-     *   Such messages can be used to provide ancillary information
-     *   not suitable for the kLogCategory_Error and
-     *   kLogCategory_Progress categories.
-     *
-     */
-    kLogCategory_Detail = 3,
-
-    /*!<
-     *   Indicates a category of log message that describes information
-     *   needed by IE and QA teams for automated testing.
-     *
-     */
-    kLogCategory_Retain = 4,
-
-    kLogCategory_Max = kLogCategory_Retain
-};
-
-/**
- * Log, to the platform-specified mechanism, the specified log
- * message, @a msg, for the specified module, @a module, in the
- * provided category, @a category.
- *
- * @param[in] module    A LogModule enumeration indicating the
- *                      source of the chip package module that
- *                      generated the log message. This must be
- *                      translated within the function to a module
- *                      name for inclusion in the log message.
- * @param[in] category  A LogCategory enumeration indicating the
- *                      category of the log message. The category
- *                      may be filtered in or out if
- *                      CHIP_LOG_FILTERING was asserted.
- * @param[in] msg       A pointer to a NULL-terminated C string with
- *                      C Standard Library-style format specifiers
- *                      containing the log message to be formatted and
- *                      logged.
- * @param[in] v         A variadic argument list whose elements should
- *                      correspond to the format specifiers in @a msg.
- *
- */
 void LogV(uint8_t module, uint8_t category, const char * msg, va_list args);
-
 void Log(uint8_t module, uint8_t category, const char * msg, ...);
+
 uint8_t GetLogFilter();
 void SetLogFilter(uint8_t category);
