+++ src/platform/BUILD.gn
@@ -241,13 +241,13 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "cc13x2_26x2/ConnectivityManagerImpl.cpp",
         "cc13x2_26x2/ConnectivityManagerImpl.h",
         "cc13x2_26x2/InetPlatformConfig.h",
-        "cc13x2_26x2/Logging.cpp",
         "cc13x2_26x2/PlatformManagerImpl.cpp",
         "cc13x2_26x2/PlatformManagerImpl.h",
         "cc13x2_26x2/SystemPlatformConfig.h",
       ]
       if (chip_enable_openthread) {
         public_deps += [
+          "${chip_root}/src/platform/logging:cc13x2_26x2",
           "${chip_root}/third_party/ti_simplelink_sdk:mbedtls",
           "${chip_root}/third_party/ti_simplelink_sdk:ti_simplelink_sdk",
           "${chip_root}/third_party/ti_simplelink_sdk:ti_simplelink_sysconfig",
@@ -273,7 +273,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "Darwin/ConnectivityManagerImpl.cpp",
         "Darwin/ConnectivityManagerImpl.h",
         "Darwin/InetPlatformConfig.h",
-        "Darwin/Logging.cpp",
         "Darwin/PlatformManagerImpl.cpp",
         "Darwin/PlatformManagerImpl.h",
         "Darwin/PosixConfig.cpp",
@@ -293,6 +292,8 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
           "Darwin/UUIDHelperImpl.mm",
         ]
       }
+
+      public_deps += [ "${chip_root}/src/platform/logging:darwin" ]
     } else if (chip_device_platform == "efr32") {
       sources += [
         "EFR32/BLEManagerImpl.cpp",
@@ -310,7 +311,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "EFR32/EFR32Config.cpp",
         "EFR32/EFR32Config.h",
         "EFR32/InetPlatformConfig.h",
-        "EFR32/Logging.cpp",
         "EFR32/PlatformManagerImpl.cpp",
         "EFR32/PlatformManagerImpl.h",
         "EFR32/SystemPlatformConfig.h",
@@ -326,7 +326,10 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "$dir_pw_kvs:crc16",
         "$dir_pw_log",
       ]
-      public_deps += [ "$dir_pw_kvs" ]
+      public_deps += [
+        "$dir_pw_kvs",
+        "${chip_root}/src/platform/logging:efr32",
+      ]
       sources += [
         "EFR32/KeyValueStoreManagerImpl.cpp",
         "EFR32/KeyValueStoreManagerImpl.h",
@@ -358,7 +361,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "ESP32/ESP32Utils.h",
         "ESP32/KeyValueStoreManagerImpl.cpp",
         "ESP32/KeyValueStoreManagerImpl.h",
-        "ESP32/Logging.cpp",
         "ESP32/LwIPCoreLock.cpp",
         "ESP32/MdnsImpl.cpp",
         "ESP32/NetworkProvisioningServerImpl.h",
@@ -373,7 +375,10 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "FreeRTOS/SystemTimeSupport.cpp",
       ]
 
-      public_deps += [ "${chip_root}/src/crypto" ]
+      public_deps += [
+        "${chip_root}/src/crypto",
+        "${chip_root}/src/platform/logging:esp32",
+      ]
     } else if (chip_device_platform == "k32w") {
       sources += [
         "FreeRTOS/SystemTimeSupport.cpp",
@@ -389,7 +394,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "K32W/DeviceNetworkProvisioningDelegateImpl.h",
         "K32W/K32WConfig.cpp",
         "K32W/K32WConfig.h",
-        "K32W/Logging.cpp",
         "K32W/NetworkProvisioningServerImpl.h",
         "K32W/PlatformManagerImpl.cpp",
         "K32W/PlatformManagerImpl.h",
@@ -407,7 +411,10 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         ]
       }
 
-      public_deps += [ "${chip_root}/src/crypto" ]
+      public_deps += [
+        "${chip_root}/src/crypto",
+        "${chip_root}/src/platform/logging:k32w",
+      ]
     } else if (chip_device_platform == "linux") {
       sources += [
         "Linux/BLEManagerImpl.cpp",
@@ -431,7 +438,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "Linux/InetPlatformConfig.h",
         "Linux/KeyValueStoreManagerImpl.cpp",
         "Linux/KeyValueStoreManagerImpl.h",
-        "Linux/Logging.cpp",
         "Linux/PlatformManagerImpl.cpp",
         "Linux/PlatformManagerImpl.h",
         "Linux/PosixConfig.cpp",
@@ -465,12 +471,14 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         public_deps += [ "Linux/dbus/bluez" ]
       }
 
-      public_deps += [ "${chip_root}/third_party/inipp" ]
+      public_deps += [
+        "${chip_root}/src/platform/logging:linux",
+        "${chip_root}/third_party/inipp",
+      ]
     } else if (chip_device_platform == "nrfconnect") {
       sources += [
         "Zephyr/BLEManagerImpl.cpp",
         "Zephyr/ConfigurationManagerImpl.cpp",
-        "Zephyr/Logging.cpp",
         "Zephyr/PlatformManagerImpl.cpp",
         "Zephyr/SystemTimeSupport.cpp",
         "Zephyr/ZephyrConfig.cpp",
@@ -498,6 +506,7 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
           "nrfconnect/ThreadStackManagerImpl.h",
         ]
       }
+      public_deps += [ "${chip_root}/src/platform/logging:zephyr" ]
     } else if (chip_device_platform == "qpg6100") {
       sources += [
         "FreeRTOS/SystemTimeSupport.cpp",
@@ -514,7 +523,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "qpg6100/DeviceNetworkProvisioningDelegateImpl.cpp",
         "qpg6100/DeviceNetworkProvisioningDelegateImpl.h",
         "qpg6100/InetPlatformConfig.h",
-        "qpg6100/Logging.cpp",
         "qpg6100/PlatformManagerImpl.cpp",
         "qpg6100/PlatformManagerImpl.h",
         "qpg6100/SystemPlatformConfig.h",
@@ -529,6 +537,7 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
           "qpg6100/ThreadStackManagerImpl.h",
         ]
       }
+      public_deps += [ "${chip_root}/src/platform/logging:qpg6100" ]
     }
 
     allow_circular_includes_from = [ "${chip_root}/src/lib/support" ]
