+++ src/platform/logging/impl/darwin/Logging.cpp
